auth: silent-only token getter; lift Popcorn/AuthError to (authed) route group
The popup fallback inside `tokenGetter` couldn't fire from a non-user-
gesture context (acquireTokenSilent is invoked from inside fetch
wrappers). When silent refresh failed during long sequences -- Moorhen
panel mounts, R-group decomposition on Materia's AggregationPage -- the
popup was browser-blocked, the request went out tokenless, and the user
saw "Your session has expired" snackbars. Replaced the popup fallback
with a 250ms retry of acquireTokenSilent, which catches the common case
of parallel fetches racing the underlying MSAL refresh. On real silent
failure we now return null and let the api-fetch layer's existing
AUTH_ERROR_EVENT path drive the snackbar + auto-redirect to sign-in.

Separately, PopcornProvider + AuthErrorHandler were nested inside
CCP4i2App, which meant they only covered routes wrapped in the app
shell. Modals/dialogs that mounted outside the shell -- and Materia's
compounds-side routes that compose CCP4i2 in -- got no popcorn surface
and 401s went into the void.

Lifted both to a new (authed)/ route group at app/ccp4i2/(authed)/.
The existing /ccp4i2/config route stays one level up, outside (authed),
because config must remain reachable when Django isn't running yet
(initial setup, server reconfiguration). RequireAuth on a non-running
backend just spins; AuthErrorHandler on a non-running backend would
surface "session expired" snackbars in response to plain connection
failures.

Added @/* path alias to renderer/tsconfig.json and migrated every
import in the moved files. URL paths are unchanged -- route groups are
invisible to the Next.js router.
