+++ client/renderer/app/ccp4i2/(authed)/job/[jobid]/page.tsx
@@ -1,0 +1,9 @@
+"use client";
+import { JobView } from "@/components/job-view";
+import { useParams } from "next/navigation";
+
+export default function JobPage() {
+  const params = useParams();
+  const { jobid } = (params ?? {}) as { jobid: string };
+  return <JobView jobid={parseInt(jobid)} />;
+}
