+++ client/renderer/app/ccp4i2/(authed)/project/layout.tsx
@@ -1,0 +1,12 @@
+"use client";
+import { PropsWithChildren } from "react";
+import { RunningProcessesProvider } from "@/providers/running-processes";
+
+export default function ProjectLayout(props: PropsWithChildren) {
+  return (
+    <RunningProcessesProvider>
+      {/* Children components will be rendered here */}
+      {props.children}
+    </RunningProcessesProvider>
+  );
+}
