--- client/renderer/app/ccp4i2/job/[jobid]/page.tsx
@@ -1,9 +1,0 @@
-"use client";
-import { JobView } from "../../../../components/job-view";
-import { useParams } from "next/navigation";
-
-export default function JobPage() {
-  const params = useParams();
-  const { jobid } = (params ?? {}) as { jobid: string };
-  return <JobView jobid={parseInt(jobid)} />;
-}
