+++ client/renderer/app/ccp4i2/layout.tsx
@@ -1,34 +1,23 @@
 "use client";
 import { PropsWithChildren } from "react";
-import { DeleteDialogProvider } from "../../providers/delete-dialog";
-import { FindInPageProvider } from "../../providers/find-in-page-provider";
-import { RecentlyStartedJobsProvider } from "../../providers/recently-started-jobs-context";
-import { CCP4i2App } from "../../providers/ccp4i2-app";
-import RequireAuth from "../../components/require-auth";
-
-const REQUIRE_AUTH = process.env.NEXT_PUBLIC_REQUIRE_AUTH === "true";
 
 /**
- * CCP4i2 app layout - wraps all ccp4i2 pages with:
- * - Authentication (when required) - AuthProvider is in root layout
- * - Job tracking context
- * - Delete dialog provider
- * - CCP4i2 app shell (header, navigation)
+ * /ccp4i2 segment layout — passthrough.
+ *
+ * The real provider stack (RequireAuth, CCP4i2App, FindInPageProvider,
+ * RecentlyStartedJobsProvider, DeleteDialogProvider, PopcornProvider,
+ * AuthErrorHandler) lives one level deeper at app/ccp4i2/(authed)/layout.tsx.
+ * Everything except /ccp4i2/config sits inside that route group.
+ *
+ * Config stays here, outside the (authed) group, because it must remain
+ * reachable when Django isn't running yet (initial setup, server
+ * reconfiguration). RequireAuth on a non-running backend just spins;
+ * AuthErrorHandler on a non-running backend would surface "session
+ * expired" snackbars in response to connection failures.
+ *
+ * Kept as an explicit file (rather than deleted) so this rationale lives
+ * with the layout boundary it defends.
  */
 export default function CCP4i2Layout(props: PropsWithChildren) {
-  return (
-    <FindInPageProvider>
-      <RecentlyStartedJobsProvider>
-        <DeleteDialogProvider>
-          {REQUIRE_AUTH ? (
-            <RequireAuth>
-              <CCP4i2App>{props.children}</CCP4i2App>
-            </RequireAuth>
-          ) : (
-            <CCP4i2App>{props.children}</CCP4i2App>
-          )}
-        </DeleteDialogProvider>
-      </RecentlyStartedJobsProvider>
-    </FindInPageProvider>
-  );
+  return <>{props.children}</>;
 }
