--- client/renderer/app/ccp4i2/moorhen-page/moorhen-page-client.tsx
@@ -1,12 +1,0 @@
-"use client";
-
-import MoorhenLoader from "../../../components/moorhen/client-side-moorhen-loader";
-import { ClientStoreProvider } from "../../../providers/client-store-provider";
-
-export default function MoorhenPageClient() {
-  return (
-    <ClientStoreProvider>
-      <MoorhenLoader fileIds={[]} />
-    </ClientStoreProvider>
-  );
-}
