--- client/renderer/app/ccp4i2/page.tsx
@@ -1,85 +1,0 @@
-"use client";
-import { Badge, Box, Container, Skeleton, Stack } from "@mui/material";
-import ProjectsToolbar from "../../components/projects-toolbar";
-import ProjectsTable from "../../components/projects-table";
-import CCP4i2TopBar from "../../components/ccp4i2-topbar";
-import { useApi } from "../../api";
-import { Project } from "../../types/models";
-
-export default function ProjectsPage() {
-  const api = useApi();
-  // Note: second param is refreshInterval, not timeout. Set to 0 to disable auto-refresh.
-  const { data: projects } = api.get<Project[]>("projects");
-
-  return (
-    <Box
-      sx={{
-        display: "flex",
-        flexDirection: "column",
-        height: "100vh",
-        overflow: "hidden",
-      }}
-    >
-      <CCP4i2TopBar title="CCP4i2 Projects" />
-      <Container
-        sx={{
-          flex: 1,
-          display: "flex",
-          flexDirection: "column",
-          overflow: "hidden",
-          py: 2,
-        }}
-      >
-        <Stack spacing={2} sx={{ flex: 1, overflow: "hidden" }}>
-          <ProjectsToolbar />
-          {projects ? (
-            projects.length > 0 ? (
-              <Box sx={{ flex: 1, overflow: "hidden" }}>
-                <ProjectsTable />
-              </Box>
-            ) : (
-              <Stack alignItems="center" spacing={2} sx={{ py: 8 }}>
-                <Badge color="secondary" badgeContent={0}>
-                  <span style={{ fontSize: 18, color: "#888" }}>
-                    No projects found
-                  </span>
-                </Badge>
-                <span style={{ color: "#666" }}>
-                  <svg width="64" height="64" viewBox="0 0 24 24" fill="none">
-                    <rect
-                      x="3"
-                      y="7"
-                      width="18"
-                      height="13"
-                      rx="2"
-                      fill="#e0e0e0"
-                    />
-                    <path
-                      d="M3 7V5a2 2 0 0 1 2-2h3.17a2 2 0 0 1 1.41.59l1.83 1.83H19a2 2 0 0 1 2 2v1"
-                      stroke="#bdbdbd"
-                      strokeWidth="2"
-                      strokeLinecap="round"
-                      strokeLinejoin="round"
-                    />
-                  </svg>
-                </span>
-                <span
-                  style={{
-                    color: "#666",
-                    fontSize: 16,
-                    textAlign: "center",
-                  }}
-                >
-                  Use the <b>New</b> or <b>Import</b> buttons above to create or
-                  load a project.
-                </span>
-              </Stack>
-            )
-          ) : (
-            <Skeleton variant="rectangular" width="100%" height={500} />
-          )}
-        </Stack>
-      </Container>
-    </Box>
-  );
-}
