--- client/renderer/app/ccp4i2/project/[id]/job/[jobid]/page.tsx
@@ -1,9 +1,0 @@
-"use client";
-import { useParams } from "next/dist/client/components/navigation";
-import { JobView } from "../../../../../../components/job-view";
-
-export default function JobPage() {
-  const params = useParams();
-  const { jobid } = (params ?? {}) as { jobid: string };
-  return <JobView jobid={parseInt(jobid)} />;
-}
