--- client/renderer/app/ccp4i2/project/layout.tsx
@@ -1,12 +1,0 @@
-"use client";
-import { PropsWithChildren } from "react";
-import { RunningProcessesProvider } from "../../../providers/running-processes";
-
-export default function ProjectLayout(props: PropsWithChildren) {
-  return (
-    <RunningProcessesProvider>
-      {/* Children components will be rendered here */}
-      {props.children}
-    </RunningProcessesProvider>
-  );
-}
